+++ micropip/package_manager.py
@@ -37,6 +37,9 @@ class PackageManager:
 
         self.index_urls = package_index.DEFAULT_INDEX_URLS[:]
         self.compat_layer: type[CompatibilityLayer] = compat
+        self.constraints: list[str] = []
+
+        pass
 
     async def install(
         self,
@@ -47,6 +50,7 @@ class PackageManager:
         pre: bool = False,
         index_urls: list[str] | str | None = None,
         *,
+        constraints: list[str] | None = None,
         verbose: bool | int | None = None,
     ) -> None:
         """Install the given package and all of its dependencies.
@@ -131,6 +135,14 @@ class PackageManager:
             - If a list of URLs is provided, micropip will try each URL in order until
             it finds a package. If no package is found, an error will be raised.
 
+        constraints :
+
+            A list of requirements with versions/URLs which will be used only if
+            needed by any ``requirements``.
+
+            Unlike ``requirements``, the package name _must_ be provided in the
+            PEP-508 format e.g. ``pkgname@https://...``.
+
         verbose :
             Print more information about the process. By default, micropip does not
             change logger level. Setting ``verbose=True`` will print similar
@@ -141,11 +153,14 @@ class PackageManager:
             if index_urls is None:
                 index_urls = self.index_urls
 
+            if constraints is None:
+                constraints = self.constraints
+
             ctx = default_environment()
             if isinstance(requirements, str):
                 requirements = [requirements]
 
-            fetch_kwargs = dict()
+            fetch_kwargs = {}
 
             if credentials:
                 fetch_kwargs["credentials"] = credentials
@@ -166,59 +181,48 @@ class PackageManager:
                 fetch_kwargs=fetch_kwargs,
                 verbose=verbose,
                 index_urls=index_urls,
+                constraints=constraints,
             )
             await transaction.gather_requirements(requirements)
 
             if transaction.failed:
-                failed_requirements = ", ".join(
-                    [f"'{req}'" for req in transaction.failed]
-                )
+                failed_requirements = ", ".join([f"'{req}'" for req in transaction.failed])
                 raise ValueError(
                     f"Can't find a pure Python 3 wheel for: {failed_requirements}\n"
                     f"See: {FAQ_URLS['cant_find_wheel']}\n"
                 )
 
-            package_names = [pkg.name for pkg in transaction.pyodide_packages] + [
-                pkg.name for pkg in transaction.wheels
-            ]
+            pyodide_packages, wheels = transaction.pyodide_packages, transaction.wheels
+
+            package_names = [pkg.name for pkg in wheels + pyodide_packages]
 
             logger.debug(
                 "Installing packages %r and wheels %r ",
                 transaction.pyodide_packages,
                 [w.filename for w in transaction.wheels],
             )
-            if package_names:
-                logger.info(
-                    "Installing collected packages: %s", ", ".join(package_names)
-                )
 
-            wheel_promises: list[Coroutine[Any, Any, None] | asyncio.Task[Any]] = []
+            if package_names:
+                logger.info("Installing collected packages: %s", ", ".join(package_names))
+
+            # Install PyPI packages
+            # detect whether the wheel metadata is from PyPI or from custom location
+            # wheel metadata from PyPI has SHA256 checksum digest.
+            await asyncio.gather(*(wheel.install(wheel_base) for wheel in wheels))
+
             # Install built-in packages
-            pyodide_packages = transaction.pyodide_packages
-            if len(pyodide_packages):
+            if pyodide_packages:
                 # Note: branch never happens in out-of-browser testing because in
                 # that case REPODATA_PACKAGES is empty.
-                wheel_promises.append(
-                    asyncio.ensure_future(
-                        self.compat_layer.loadPackage(
-                            self.compat_layer.to_js(
-                                [name for [name, _, _] in pyodide_packages]
-                            )
+                await asyncio.ensure_future(
+                    self.compat_layer.loadPackage(
+                        self.compat_layer.to_js(
+                            [name for [name, _, _] in pyodide_packages]
                         )
                     )
                 )
 
-            # Now install PyPI packages
-            for wheel in transaction.wheels:
-                # detect whether the wheel metadata is from PyPI or from custom location
-                # wheel metadata from PyPI has SHA256 checksum digest.
-                wheel_promises.append(wheel.install(wheel_base))
-
-            await asyncio.gather(*wheel_promises)
-
-            packages = [
-                f"{pkg.name}-{pkg.version}" for pkg in transaction.pyodide_packages
-            ] + [f"{pkg.name}-{pkg.version}" for pkg in transaction.wheels]
+            packages = [f"{pkg.name}-{pkg.version}" for pkg in pyodide_packages + wheels]
 
             if packages:
                 logger.info("Successfully installed %s", ", ".join(packages))
@@ -504,3 +508,16 @@ class PackageManager:
             urls = [urls]
 
         self.index_urls = urls[:]
+
+    def set_constraints(self, constraints: List[str]):  # noqa: UP006
+        """
+        Set the default constraints to use when looking up packages.
+
+        Parameters
+        ----------
+        constraints
+            A list of PEP-508 requirements, each of which must include a name and
+            version, but no ``[extras]``.
+        """
+
+        self.constraints = constraints[:]