+++ micropip/wheelinfo.py
@@ -7,10 +7,6 @@ from pathlib import Path
 from typing import Any, Literal
 from urllib.parse import ParseResult, urlparse
 
-from packaging.requirements import Requirement
-from packaging.tags import Tag
-from packaging.version import Version
-
 from ._compat import (
     fetch_bytes,
     get_dynlibs,
@@ -18,7 +14,11 @@ from ._compat import (
     loadedPackages,
 )
 from ._utils import parse_wheel_filename
+from ._vendored.packaging.src.packaging.requirements import Requirement
+from ._vendored.packaging.src.packaging.tags import Tag
+from ._vendored.packaging.src.packaging.version import Version
 from .metadata import Metadata, safe_name, wheel_dist_info_dir
+from .types import DistributionMetadata
 
 
 @dataclass
@@ -43,6 +43,7 @@ class WheelInfo:
     parsed_url: ParseResult
     sha256: str | None = None
     size: int | None = None  # Size in bytes, if available (PEP 700)
+    core_metadata: DistributionMetadata = None  # Wheel's metadata (PEP 658 / PEP-714)
 
     # Fields below are only available after downloading the wheel, i.e. after calling `download()`.
 
@@ -50,7 +51,8 @@ class WheelInfo:
     _metadata: Metadata | None = None  # Wheel metadata.
     _requires: list[Requirement] | None = None  # List of requirements.
 
-    # Path to the .dist-info directory. This is only available after extracting the wheel, i.e. after calling `extract()`.
+    # Path to the .dist-info directory.
+    # This is only available after extracting the wheel, i.e. after calling `extract()`.
     _dist_info: Path | None = None
 
     def __post_init__(self):
@@ -58,6 +60,7 @@ class WheelInfo:
             self.url.startwith(p) for p in ("http:", "https:", "emfs:", "file:")
         ), self.url
         self._project_name = safe_name(self.name)
+        self.metadata_url = self.url + ".metadata"
 
     @classmethod
     def from_url(cls, url: str) -> "WheelInfo":
@@ -66,8 +69,15 @@ class WheelInfo:
         See https://www.python.org/dev/peps/pep-0427/#file-name-convention
         """
         parsed_url = urlparse(url)
-        if parsed_url.scheme == "":
-            url = "file:///" + url
+        if not parsed_url.scheme:
+            parsed_url = ParseResult(
+                scheme="file",
+                netloc=parsed_url.netloc,
+                path=parsed_url.path,
+                params=parsed_url.params,
+                query=parsed_url.query,
+                fragment=parsed_url.fragment,
+            )
         file_name = Path(parsed_url.path).name
         name, version, build, tags = parse_wheel_filename(file_name)
         return WheelInfo(
@@ -89,6 +99,7 @@ class WheelInfo:
         version: Version,
         sha256: str | None,
         size: int | None,
+        core_metadata: DistributionMetadata = None,
     ) -> "WheelInfo":
         """Extract available metadata from response received from package index"""
         parsed_url = urlparse(url)
@@ -104,6 +115,7 @@ class WheelInfo:
             parsed_url=parsed_url,
             sha256=sha256,
             size=size,
+            core_metadata=core_metadata,
         )
 
     async def install(self, target: Path) -> None:
@@ -130,10 +142,42 @@ class WheelInfo:
         if self._data is not None:
             return
 
-        self._data = await self._fetch_bytes(fetch_kwargs)
-        with zipfile.ZipFile(io.BytesIO(self._data)) as zf:
-            metadata_path = wheel_dist_info_dir(zf, self.name) + "/" + Metadata.PKG_INFO
-            self._metadata = Metadata(zipfile.Path(zf, metadata_path))
+        self._data = await self._fetch_bytes(self.url, fetch_kwargs)
+
+        # The wheel's metadata might be downloaded separately from the wheel itself.
+        # If it is not downloaded yet or if the metadata is not available, extract it from the wheel.
+        if self._metadata is None:
+            with zipfile.ZipFile(io.BytesIO(self._data)) as zf:
+                metadata_path = (
+                    Path(wheel_dist_info_dir(zf, self.name)) / Metadata.PKG_INFO
+                )
+                self._metadata = Metadata(zipfile.Path(zf, str(metadata_path)))
+
+    def pep658_metadata_available(self) -> bool:
+        """
+        Check if the wheel's metadata is exposed via PEP 658.
+        """
+        return self.core_metadata is not None
+
+    async def download_pep658_metadata(
+        self,
+        fetch_kwargs: dict[str, Any],
+    ) -> None:
+        """
+        Download the wheel's metadata. If the metadata is not available, return None.
+        """
+        if self.core_metadata is None:
+            return None
+
+        data = await self._fetch_bytes(self.metadata_url, fetch_kwargs)
+
+        match self.core_metadata:
+            case {"sha256": checksum}:  # sha256 checksum available
+                _validate_sha256_checksum(data, checksum)
+            case _:  # no checksum available
+                pass
+
+        self._metadata = Metadata(data)
 
     def requires(self, extras: set[str]) -> list[Requirement]:
         """
@@ -148,14 +192,14 @@ class WheelInfo:
         self._requires = requires
         return requires
 
-    async def _fetch_bytes(self, fetch_kwargs: dict[str, Any]):
+    async def _fetch_bytes(self, url: str, fetch_kwargs: dict[str, Any]):
         if self.parsed_url.scheme not in ("https", "http", "emfs", "file"):
             # Don't raise ValueError it gets swallowed
             raise TypeError(
-                f"Cannot download from a non-remote location: {self.url!r} ({self.parsed_url!r})"
+                f"Cannot download from a non-remote location: {url!r} ({self.parsed_url!r})"
             )
         try:
-            bytes = await fetch_bytes(self.url, fetch_kwargs)
+            bytes = await fetch_bytes(url, fetch_kwargs)
             return bytes
         except OSError as e:
             if self.parsed_url.hostname in [
@@ -165,7 +209,7 @@ class WheelInfo:
                 raise e
             else:
                 raise ValueError(
-                    f"Can't fetch wheel from '{self.url}'. "
+                    f"Can't fetch wheel from {url!r}. "
                     "One common reason for this is when the server blocks "
                     "Cross-Origin Resource Sharing (CORS). "
                     "Check if the server is sending the correct 'Access-Control-Allow-Origin' header."